Rearrange deleteprogram
